flare

+++ flare_module/test/command/gatling.py
@@ -40,33 +40,40 @@ class GatlingServer():
         with cd(FlareResult.REMOTE_GATLING_HOME):                                       # cd gatling home
             self.execute(self.MVN_TEST.format(simulationClass, outputBaseName, jvm))    # run gatling
 
-
-    def result_download(self):
+    def remove_old_report(self):
         with cd(FlareResult.REMOTE_GATLING_RESULT):
             lsOutput = run('ls')
             fileNames = lsOutput.split()
             for fileName in fileNames:
-                file = fileName.split('-')
-                timestamp = file[2]
-                timestamp = datetime.fromtimestamp(int(timestamp) / 1000).strftime('%Y%m%d%H%M%S')
-                date = timestamp[:8]
-                downloadPath = self.FLARE_RESULT_GATLING + '/' + date
-                changeFilename = file[0] + '-' + file[1] + '-' + timestamp
+                run('rm -rf {0}*'.format(fileName))
 
-                run('tar -cf {0}.tar {0}'.format(fileName))                 # tar gatling report
-                local('mkdir -p {0}'.format(downloadPath))                  # mkdir download path
-                get('{0}.tar'.format(fileName), downloadPath)               # download gatling.tar
-                run('rm -rf {0}*'.format(fileName))                         # remove gatling report
+    def result_download(self, simulationClassName, outputBaseName):
+        with cd(FlareResult.REMOTE_GATLING_RESULT):
+            lsOutput = run('ls')
+            fileNames = lsOutput.split()
+            for fileName in fileNames:
+                if fileName in simulationClassName:
+                    file = fileName.split('-')
+                    timestamp = file[1]
+                    timestamp = datetime.fromtimestamp(int(timestamp) / 1000).strftime('%Y%m%d%H%M%S')
+                    date = timestamp[:8]
+                    downloadPath = self.FLARE_RESULT_GATLING + '/' + date
+                    changeFilename = outputBaseName + '-' + timestamp
 
-                with lcd(downloadPath):
-                    local('tar -xf {0}.tar'.format(fileName))               # tar gatling report
-                    local('rm -rf {0}.tar'.format(fileName))                # remove tar file
-                    local('mv {0} {1}'.format(fileName, changeFilename))    # change name
+                    run('tar -cf {0}.tar {0}'.format(fileName))                 # tar gatling report
+                    local('mkdir -p {0}'.format(downloadPath))                  # mkdir download path
+                    get('{0}.tar'.format(fileName), downloadPath)               # download gatling.tar
+                    run('rm -rf {0}*'.format(fileName))                         # remove gatling report
 
-                    last_log_info = FlarePath.FLARE_RESULT + '/last_log_info'
-                    f = open(last_log_info, 'w', encoding='UTF8')
-                    f.write(downloadPath + '/' + changeFilename)
-                    f.close()
+                    with lcd(downloadPath):
+                        local('tar -xf {0}.tar'.format(fileName))               # tar gatling report
+                        local('rm -rf {0}.tar'.format(fileName))                # remove tar file
+                        local('mv {0} {1}'.format(fileName, changeFilename))    # change name
+
+                        last_log_info = FlarePath.FLARE_RESULT + '/last_log_info'
+                        f = open(last_log_info, 'w', encoding='UTF8')
+                        f.write(downloadPath + '/' + changeFilename)
+                        f.close()
